+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [7.1.0] - 2023-02-18
+
 ### Added 
 
 * Authentication package, with a Laravel Fortify action for rehashing user's password after login.
@@ -1026,7 +1028,8 @@ It will highjack the `app` binding, which will cause your application to behave 
 
 * Please review commits on [GitHub](https://github.com/aedart/athenaeum/commits/master)
 
-[Unreleased]: https://github.com/aedart/athenaeum/compare/7.0.1...HEAD
+[Unreleased]: https://github.com/aedart/athenaeum/compare/7.1.0...HEAD
+[7.1.0]: https://github.com/aedart/athenaeum/compare/7.0.1...7.1.0
 [7.0.1]: https://github.com/aedart/athenaeum/compare/7.0.0...7.0.1
 [7.0.0]: https://github.com/aedart/athenaeum/compare/7.0.0-alpha.1...7.0.0
 [7.0.0-alpha.1]: https://github.com/aedart/athenaeum/compare/7.0.0-alpha...7.0.0-alpha.1